requirements.txt
anyio==3.6.2
autopep8==2.0.1
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.6
pycodestyle==2.10.0
pydantic==1.10.5
python-dotenv==0.21.1
python-multipart==0.0.5
python-nmap==0.7.1
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
faiss-cpu==1.7.3
fqdn==1.5.1
gensim==4.3.1
isoduration==20.11.0
jsonpointer==2.3
jupyter==1.0.0
nltk==3.8.1
pip-chill==1.0.1
pyflann==1.6.14
pysocks==1.7.1
selenium==4.8.3
sqlalchemy==2.0.7
torch==2.0.0
transformers==4.27.3
uri-template==1.2.0
vectorhub==1.8.3
virtualenv==20.21.0
webcolors==1.12
